requirements.txt
#  Web Framework 
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9

#  LLM & Embeddings 
langchain==0.3.0
langchain-community==0.3.0
langchain-ollama==0.2.0
langchain-openai==0.2.0
openai==1.51.0
ollama==0.3.3

#  Vector Store 
chromadb==0.5.11
faiss-cpu==1.8.0

#  Document Loaders 
pypdf==4.3.1
python-docx==1.1.2
unstructured==0.15.13

#  UI 
gradio==4.44.0

#  Config & Env 
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
pyyaml==6.0.2

#  Utilities 
httpx==0.27.2
tenacity==9.0.0
tqdm==4.66.5
numpy==1.26.4

#  Training 
torch==2.4.1
transformers==4.44.2
datasets==3.0.1
peft==0.12.0
accelerate==0.34.2
evaluate==0.4.3
